requirements.txt
amqp==5.2.0       
asgiref==3.7.2       
async-timeout==4.0.3       
attrs==23.2.0      
backports.zoneinfo==0.2.1       
billiard==4.2.0       
celery==5.4.0       
certifi==2024.2.2    
charset-normalizer==3.3.2       
click==8.1.7       
click-didyoumean==0.3.0       
click-plugins==1.1.1       
click-repl==0.3.0       
Django==4.1.13      
django-celery-results==2.5.1       
django-extensions==3.2.3       
djangorestframework==3.14.0      
djongo==1.3.6       
dnspython==2.4.2       
exceptiongroup==1.2.1
flower==2.0.1
h11==0.14.0
humanize==4.9.0
idna==3.7
kombu== 5.3.4
outcome==1.3.0.post0
pika==1.3.2
pip==20.0.2
pkg-resources==0.0.0
prometheus-client==0.20.0
prompt-toolkit==3.0.41
PyJWT== 2.8.0
pymongo== 3.12.3
PySocks== 1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
rapidfuzz==3.6.1
redis== 5.0.1
requests==2.31.0
selenium==4.21.0
setuptools==44.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.2.4
tornado==6.4
trio==0.25.1
trio-websocket==0.11.1
typing-extensions==4.12.0
tzdata==2023.3
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.12
wheel==0.43.0  
wsproto==1.2.0